Add ability for child component to change parent component's state
