+++ src/Components/main.js
@@ -5,10 +5,10 @@ import Typography from '@material-ui/core/Typography';
 
 // Dependencies created by our group
 import { mainStyles } from '../styles/styles';
-import { getLocation } from '../location/getLocation';
 import { HourGraph } from './Graphs/Hourgraph';
-import { ForecastGraph } from './Graphs/Forecastgraph';
+import { ForecastGraph } from './Graphs/ForecastGraph';
 import { CurrentWeatherDisplay} from './CurrentWeatherDisplay/CurrentWeatherDisplay';
+import { Locator } from './Locator/Locator';
 
 /* One way of styling with Material UI is to use hooks.
    However, these only work inside functions, so we can't have this with
@@ -32,7 +32,8 @@ export class Main extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
-            latitude,
+            latitude: 0,
+            longitude: 0,
             temperature: 50,
             uvIndex: 10,
             humidity: 60,
@@ -42,26 +43,14 @@ export class Main extends React.Component {
             hourlyData: {},
             units: "F"
         };
-        this.handleDebugClick = this.handleDebugClick.bind(this);
         this.handleAPIClick = this.handleAPIClick.bind(this);
+        this.stateHandler = this.stateHandler.bind(this);
     }
 
-    // Kind of temporary, for debug purposes and to demonstrate the location feature.
-    async handleDebugClick(e) {
-        e.preventDefault();
-        let result;
-        result = await getLocation();
-        if (result.success) {
-            console.log(result);
-            this.setState({
-                city: result.name,
-                country: result.sys.country,
-                temperature: result.main.temp
-            });
-        }
-        else {
-            console.log("DEBUG: Could not find your location.");
-        }
+
+
+    stateHandler(values) {
+        this.setState(values);
     }
 
     handleAPIClick(e) {
@@ -82,9 +71,9 @@ export class Main extends React.Component {
                 <HeaderHook />
                 <HourGraph />
                 <ForecastGraph />
-                <Button variant="contained" color="primary" onClick={this.handleDebugClick}>
-                    Debug - Get Location
-                </Button>
+                <Locator 
+                    stateHandler={this.stateHandler}
+                />
                 <Button variant="contained" color="primary" onClick={this.handleAPIClick}>
                     Debug - Make API Call
                 </Button>
